Update FloorOfNumber.java
Problem Statement:-  
Given a sorted array, find the "Floor" of a target number.

The "Floor" of a number is the greatest element in the array that is "less than or equal to" the target.  
If the target is present, return its index. If not, return the index of the greatest element just smaller than the target.

Approach:-  
We use the Binary Search algorithm to efficiently find the Floor.

 Binary Search Logic:-  
Set `start = 0` and `end = arr.length - 1`.

In each iteration:-

- Calculate `mid = start + (end - start) / 2`
- If `arr[mid] == target`, return `mid` (target found)
- If `arr[mid] < target`, it's a possible floor → search in right side → `start = mid + 1`
- If `arr[mid] > target`, floor must be on the left side → `end = mid - 1`

If the target is not found, return `end`.  
Why? Because `end` will be pointing to the largest element less than or equal to the target, i.e., the **floor**.

Time Complexity: - 
O(log n) – because we use Binary Search

Efficient for large sorted arrays